requirements.txt
openai>=1.108.1
python-dotenv
rich